+++ Backend/DTOs/CreateRoasterDto.cs
@@ -14,11 +14,11 @@ public class CreateRoasterDto
   [MaxLength(500, ErrorMessage = "Location Address must be at most 500 characters long.")]
   public string? LocationAddress { get; set; }
 
-  [Range(typeof(double), "-90", "90", ErrorMessage = "Latitude must be between -90 and 90.")]
-  public double? LocationCoordinateLatitude { get; set; }
+  [Range(typeof(float), "-90", "90", ErrorMessage = "Latitude must be between -90 and 90.")]
+  public float? LocationCoordinateLatitude { get; set; }
 
-  [Range(typeof(double), "-180", "180", ErrorMessage = "Latitude must be between -90 and 90.")]
-  public double? LocationCoordinateLongitude { get; set; }
+  [Range(typeof(float), "-180", "180", ErrorMessage = "Latitude must be between -90 and 90.")]
+  public float? LocationCoordinateLongitude { get; set; }
 
   [MaxLength(2000, ErrorMessage = "Description must be at most 2000 characters long.")]
   public string? Description { get; set; }
